refactor(region): move to genomics module
